requirements.txt
lightrag
openai
python-dotenv
networkx
pyvis
supabase
numpy
asyncio
